requirements.txt
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
curl_cffi==0.13.0
feedparser==6.0.12
filelock==3.19.1
frozendict==2.4.6
fsspec==2025.9.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mpmath==1.3.0
multitasking==0.0.12
narwhals==2.6.0
networkx==3.5
newsapi-python==0.2.7
nltk==3.9.2
numpy==2.3.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
patsy==1.0.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.1
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-twitter==3.5
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8
statsmodels==0.14.5
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.2
tweepy==4.16.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websockets==15.0.1
yfinance==0.2.66
